Improvement on the jerky animation
See issue #2

Simplified the animation of the invaders. Instead of doing interpolation
between ticks, we just calculate the velocity required to travel the whole
game space in screen space and then integrate over time. Also removed
the game/animation synch step.

The motion is somewhat improved but still not perfect.