Implement a simple caching audio buffer allocator
It seems that we can take a little hit from the audio buffer allocation
and experience jank in the rendering thread (imagine that) when audio
clip is played. This happens mostly on Optimus when running on NVIDIA
graphics chip (we also get less fps than when on the integrated Intel
graphics chip).

The quick solution is to cache and reuse the audio buffers instead of
continuously allocate/deallocate.