Refactored the input and menu/hud drawing
Refactored the input and menu/hud drawing behind a "state" abstraction
and separate implementation classes per each game state such as
"display help" or "display settings" or "play game".

This change simplifies the input handling logic as well as the drawing
logic since we only need to deal with a single state object (which
is at the top of state stack) at any given time.