Add a workaround for msvs failure with source files with same name
Add a workaround for msvs failure with source files with same name.
Easiest solution is to simply rename the files to have unique name.

There are two source files (translation units) with same device.cpp
name. audio/device.cpp and graphics/device.cpp. msvs 2015 fails to
deal with this and shits itself in the linking stage.