Move shader compilation into the drawable/material classes
Move shader compilation into the drawable/material classes. Shaders
can be considered to be graphics resources similar to geometry or
textures and  these details need to be managed by the drawable and
material implementations. This change then makes this graphics resource
management more consistent and allows these details to be removed from
the painter implementation.