Refactor the particle engine to be statically configurable
Refactor the particle engine to be statically configurable using
policies. We have the following policies (algorithms) that can be
used to configure the engine:
  - MotionPolicy, update the particle's position
  - UpdatePolicy, update the particle's other attributes
  - BoundsPolicy, decide what to do at particle space bounds
  - LifetimePolicy, decide what do at the end of lifetime

Each policy can update the state of the particle and return true
or false. On false the particle is going to be killed.

Additionally adding an option for respawning particles to replace
particles that have been killed.