Refactor particle engine initial particle position logic
Refactor particle engine initial particle position logic by allowing
the user to define a (sub) rectangle inside the particle space in which
the particles will begin life. This will make it possible to have for
example all particles spawn in the middle of the space and then
scather in all directions. This is useful for explosions.