Add SpriteMap material implementation
Add SpriteMap material implementation. This is similar to SpriteSet
and can be used to do simple sprite based animations by cycling
through a set of pre-rendered texture based frames at some frame rate.
The difference to SpriteSet is that instead of cycling through a set of
separate texture objects it only uses a single texture with sub-regions
within the texture being used as a single frame.