Add a simple tessellated circle drawable using triangle fan
Add a simple tessellated circle drawable using triangle fan. This
is used a fixed number of slices regardless of the final pixel
density on screen (which depends on the transformation.

TODO: Probably need to be able to derive some pixel density values
and use that to compute a LOD value for deciding in how many slices
to tesellate the circle.

Enable default construction of Materials. This is a bit unsafe but
makes life easier, for example it's possible to make have an array
of materials now.

Add a shapes rendering test in the graphics test application.