Support static/dynamic text buffers
Support static/dynamic text buffers.

Static text buffers are buffers whose content's are not expected to
change from frame to frame. They're rasterized once and uploaded to
the GPU and reused. Only if the dimensions of the text buffers have
changed (for example because of appliation window dimensions have
changed thus triggering re-layout of the HUD) the buffer contents
are re-uploaded to the GPU.

Dynamic text buffers (are unnamed) and expected to have content
that changes often/from frame to frame. These buffers are rasterized
and uploaded to the GPU every single time they're rendered.