Use correct (from font metrics) line height for rasterized text
Use correct (from font metrics) line height for rasterized text.
For now we bake this information in the rasterized bitmap of each
line of text since we don't have the lineheight info in the top level
function that lays out multipline lines of text. A slightly more
efficient implementation would propagate the line height information
up to the caller instead of baking in the rasterzized line bitmap
in order to save a few pixel ops later.