Use a hash value to identify TextBuffer content on the GPU
Use a hash value to identify TextBuffer content on the GPU. The
content of each TextBuffer can be used to compute a unique hash
which is then to used to identify the matching GPU resource i.e
texture.This means that when rendering text content that has
already been rasterized and uploaded to the GPU we can skip
re-rasterization if there already exists a texture with the
texture name matching the hash of the text buffer content.
Logically then any changes to the content of any text buffer
will produce a different hash thus triggering re-rasterization
and the creation of a new texture object.

In general this rasterization/texture creation sceheme works well
except that it's possible that content that changes often would
produce a lot of textures that are only  used once and then left
to consume space on the GPU. To mitigate that issue some kind of
"garbage collection" cycle should be implemented.