Support simplistic graphics device garbage collection
Support simplistic graphics device garbage collection. This system
is mostly designed for deleting textures that are now longer used.
A specific use case is for text rendering where text is generated
dymamically (for example HUD) and the changes trigger re-rasterization
and new texture generation. Without periodic deletion of texture
objects we'd waste device memory with texture objects that are no
longer being used.

The implementation is simply tracking the number of frames being
rendered and deleting objects which have not been used within
the last n frames, where n is value the application can pass to the
device when running the garbage cleanup cycle.