Allow bitmap copying to start at negative offsets
Allow bitmap copying to start at negative offset. This makes sense
for example when rasterizing text and copying lines of rasterized
text into target bitmap. The layout might have partially visible lines
that are offset so that they start "above" or "left of" of the target
bitmap.

If pixels that are not within the destination bitmap are clipped.