Refactor some kinks and kludges in view transforms
Refactor some kinks and kludges in view transforms. Previously there
were some hacks in computing the positions for various objects in
window space (for example for the Invader animations). This patch
fixes and removes those hacks. In particular this patch:

- Clarifies the purpose of "TransformState". Basically it is just a
  simple GridLayout that allows one to transform/map row/col
  coordinates inside a grid within the given rectangle into the
  layout's parent coordinate (pixel) space i.e. the window space
  so that the transformed coordinates are compatible with (Q)Painter.

- Removes hacks related to Invader position transforms.

- Adds a helper method that creates a GridLayout object that can
  then map coordinates in the game space into window space.

- Introduces the clarifying idea that all objects and animations related
  to game objects should use normalized game coordinates and then when
  painting/drawing the object the normalized coordinates should get
  transformed into window space coordinates.

TODO: Refactor rest of the game space animations to apply with the
above painting transformation. This requires further refactoring
of the paint methods.