Remove GridLayout from Animation interface
Remove GridLayout from Animation interface. The animation interface
is agnostic as to in which space the animated object lives. Instead
we should just provide as parameter the viewport rectangle in which
to map the animation when doing the painting. This mapping can
happen through game space transformation or by some other some means.

TODO: This change currently breaks the UFO debris animation since
the Debris a animation currently assumes only game space. This needs
to be fixed in a later commit.