Separate text styling into a separate style properties class
Separate text styling into a separate style properties class. This
makes it easier to provide API method(s) for adding text to a
TextBuffer since such API only needs to have the required parameters
and the optional properties can be set easily in the TextStyle object
that is returned from the method. Otherwise the possible combinations
of various text parameters would become complicated to handle.

This patch also does the following:

- Removes the fixed text positioning since it's currently not needed
  and clutters the API and implementation.
- Adds two new text styling properties:
  - line height scale factor
  - boolean flag for text underline

In the future it's also possible to support text with multiple colors.
This would require the following changes:
  - Use a RGBA bitmap for the rasterized text.
  - Blit/Blend the rasterized glyph into RGBA bitmap using the rasterized
    glyph 8-bit bitmap for the alpha values only. Use the text color
    property for the R, G and B values.
  - Change any material using the Bitmap texture to handle the bit
    depth appropriately and sample the texture appropriately.