Remove the Get/SetState methods from the graphics device interface
Remove the Get/SetState methods from the graphics device interface.
These were originally used as a "quick-and-dirty" type of solution
to deal with the fact that QPainter and the custom graphics device
were both using the same underlying OpenGL context and assuming
certain implicit state of the context.

Currently this behaviour is not needed and when it's needed it should
be implemented somewhere else and it should not be in the device
interface. Rather the OpenGL context state serialization and
deserizaliation is an orthogonal problem and does not need to be
part of this abstraction.