Remove dependency to Qt OpenGL from the OpenGL ES2 device
Remove dependency to Qt OpenGL from the OpenGL ES2 device.

Ideally it should be possible to  use and re-use the  device
implementation in a manner that is agnostic of any particular window
system such as GLX, EGL, WGL or AGL and that is also agnostic of any
particular OpenGL implementation. Examples of such scenarios would be
using OpenGL emulation through libANGLE or using the device with a
context that is provided by some other abstraction/framework layer
such as Qt. In either case we don't want the device to depend on
any particular implementation details. What is needed however is
a way to resolve the API entry points at runtime so that the device
can actually perform OpenGL ES2 API calls.

This patch adds a Context interface with a method for querying the
underlying implemetantation for the API function pointers. The device
upon construction with the given context will then proceed to resolve
the required function API entry points. This is the minimum required
"plumbing" to make this work.