Add initial support for editing particle systems in the editor app
Add initial support for editing particle systems in the editor app.
This commit provides all the functionality for setting the currently
possible particle system parameters.

What is missing completely is the functionality for loading existing
systems and saving/exporting the current parameters to a file.

This commit also provides re-usable GfxWidget for using the gfx
painter and device with the Qt provided GL widget.
parameters.