Refactor the audio device implementation to reduce copying
Refactor the audio buffer implementation to reduce copying of PCM
data. On windows the waveout device has to allocate buffer with
certain alignments. So a more performance minded implementation
lets the audio source (i.e the AudioSample) object write the
data directly into those buffers without any intermediate buffers.

Note that on pulseaudio we still make a temp buffer and a copy
and then call pa_stream_write with that buffer. This could still
be improved by keeping a single buffer and using a completion callback.
Regardless after this refactoring the performance on Pulseaudio
should be no worse than before and much improved on Waveout.

Because the AudioSample object now has mutable state (looped playback)
the Player API (and the device API as well) must change so that
the ownership of the AudioSample is passed down the object hierarchy
apprpropriately and only the running stream can contain the object
poiter. Nobody else should make any calls to the same AudioSample
object.