Tweak the garbage collection cleanup setting to more frames
Tweak the garbage collection cleanup setting to more frames. This
is a workaround for the problem that blinking text happens to
blink frequently enough to have the rasterized alpha mask get cleaned
away between the blink intervals.

This isn't not a solution but a quick fix. Likely the garbage
collection needs to become more sophisticated.