Fix texture binding bugs in OpenGL ES2 device
Fix texture binding bugs in OpenGL ES2 device. The "naive" code that
sets texture parameters by first selecting a texture unit and then
binding the texture to the unit and then setting a texture setting
(such as min filter) can destroy previously set state. For example
if the caller is running a loop and creating multiple textures
then setting texture properties and finally setting the texture
to a program sampler, on the next iteration of the loop the previously
set program sampler is lost if/when the next texture's properties
are being set. This is not what is wanted.

Solution is to defer setting the texture properties until the texture
is actually bound to a unit. The only exception is doing the texture
upload which we don't want to defer (because we would have to hold
on to the original image data buffer) and in that operation we must
then make sure not to overwrite the texture unit binding.