Cleanup bitmap constructor and assignment
Cleanup bitmap constructor and assignment. We actually don't need
to manually define copy ctor or move ctor. Simply make the IBitmap
copy and move ctor defined (not deleted) but only accessible to a
derived class and then in Bitmap class implicit copy and move ctor
will kick in and do the job for us. Same for the move/copy assingment.

Additionally simplify the constructor doing a strided copy.