Refactor and simplify MainWindow structure and tab management
Refactor and simplify MainWindow structure and tab management.
Simplify the tab management by letting the QTabWidget keep the tab
objects (we still need to manually delete the widgets though) and
have all tabs live on the free store instead of some living on the
stack. This simplification is possible by moving the application
event log into a (dockable) widget in the main window itself.
This frees the main tab to be used only for "content" tabs.