Add saving particle systems and their materials
Add saving particle systems and their materials. When a new particle
system is created there's no requirement to have a material available.
Instead the user can manipulate some material settings (in very
limited fashion) and when the particle system is saved a matching
material is created.

This means that subsequently it's possible to open the newly created
material in the material editor and change all of the available
material parameters. This means that there's a small subset of
parameters that the particle and material widgets must share;

  * texture files
  * texture rectangles
  * base color
  * surface type

This means that the particle widget must always first check if the
material already exists (for example when saving) and the if it does
exist then overwrite only the values that have been changed in the UI.