Add the first version of animation widget
Add the first version of animation widget. The goal of this widget
is to enable the user to compose several animation components to
a single animation where each component can be animated/moved
indepedently. Each component will have some renderable properties
such as the drawable shape in question and the material that will
be used to render it.

This is very early version and mostly just brings the barebones
widget with some placeholder UI features.