Catch std::exceptions in test_minimal's main function
Catch std::exceptions in test_minimal's main function and print to
std::cout.

An uncaught exception from main will/can cause abnormal program
termination and not much information gets shown to the user (the
developer). By catching the exceptions we can inform the developer
that the testing failed and print the contents of the exception to
hopefully help with the trouble shooting.