Support opening resources in new windows instead of main tab
Support opening resources in new windows instead of main tab. The
tab has the problem that user can only really see one resource at
a time and needs to swap between multiple tabs. This is inconvenient
when for example editing a resource that is used somewhere else (for
example an animation uses some material) and the user wants to see
the changes in the resource in the window that uses it. By opening
the resource that is being modified in a new window this problem
can be solved and it's possible to see both windows at the same time
as long as the user has enough screen space.