Simplify MainWidget interface and widget setup
Simplify MainWidget interface and widget setup. We can remove the
startup function since it's not currently needed, also instead of
having a virtual function for setting the workspace the workspace
can be passed in the constructor. This change is possible since
in the current design the widgets (and their content) is always
specific to a workspace thus the workspace cannot change during
the lifetime of a widget.