Simplify the particle system editor by removing material properties
Simplify the particle system editor by removing material properties
that were used to override the material properties. A much simpler
solution is just to have a rendering option for selecting a material
and have this default to checkerboard. This change makes the whole
UX/UI much less confusing (and also simplifies the implementation).