Draw Animation components in the order defined by the layer parameter
Draw Animation components in the order defined by the layer parameter.

This solves the problem of rendering the animation components as the
user wishes them to be rendered. However the problem is that this won't
let us apply early Z culling in the hardware. Early Z is an optimization
where the geometry is drawn from front to back thus filling the Z buffer
early. The driver can then discard the fragments by doing early Z testing
and the fragments that are covered by fragments closer to the viewer
don't need to be shaded. (The fragment shader doesn't need to run).

Note that for alpha blended content the above doesn't work but the
rendering order must be from back to front so that the blending works
correctly.