Fix gl_PointCoord sampling in texture_map shader
Fix gl_PointCoord sampling in texture_map shader. According to the
the GL ES 2.0 specification (3.3 Points) the gl_PointCoord varies
from 0 to 1 left to right on the X axis and 0 to 1 top to bottom
on the Y axis. The Y axis value is inversed wrt texture coords used
by the primitive shapes.