Apply inverse view transformation for correct object placement
Apply inverse view transformation for correct object placement. When
placing an object we must appropriately account for the view transformation
and not assume that the coordinates in the viewport/window/widget map
identially to the coordinates of the object space.

Instead map the window coordinates (through mouse events) into object
space through inverse view transformation.