Add support for simple hit test object selection in animation widget
Add support for simple hit test object selection in animation widget.
This is based on a "coarse" hit testing in animation that compares
the hit point against every animation nodes' "size" box, i.e. the box
that contains the nodes' drawable objects.

Add support for doing a hit test in the animation widget when the
mouse is clicked and selecting the object in the tree widget if any
or clearing any current selection if nothing is intersected.