Refactor the animation node transformation and object position
Refector the animation node transformation and object position so
that the object's position is now with respect to the shape's center.
I.e. the center point of the shape aligns with the position of the
node/shape.

This change arguably makes the rest of the node's transformations
more intuitive to the user. Alternative is to use the corner as
was done before but then the correlation between position is kinda
lost when the object is rotated, unless the rotation is around the
origin of the shape (it's corner) but this feels rather un-natural.

Change the animation widget tools to respect the new definition for
node's position (translation) parameter.