Fix a bug introduced in 3fc0215ff5f3e066abe1788efa32dd2da026fa32
Fix a bug introduced in 3fc0215ff5f3e066abe1788efa32dd2da026fa32.
In that change AnimationNode implementation was changed so that the
drawable shape's center would be aligned around the node's position.
However this translate was added to the node's transformation instead
of the node's drawable (model) transform. This means that the offset
would be carried over to child nodes as well, i.e. their position
0, 0, would no longer map to the center of the parent node.

This commit moves the translation from node's transformation to the
node's drawable (model) transformation.