Make mouse movement in move tool relative to node's parent for child nodes
Make mouse movement in move tool relative to node's parent for child nodes.
If the object being moved is part of a scene tree hiearchy (i.e. it
has some parent) we can't directly use the mouse movement delta in
animation space for translating the object since the parent node
might introduce a rotational component which we need to inverse.

The logic is a little bit complicated but basically if the node
has no parent (other than the empty render parent) we can directly
use mouse movement in animation space for translating the object
but if the object has a parent then we can express the translation
(and mouse delta) relative to the parent.