Flip drawable shapes' texture coordinates on Y axis
Flip drawable shape's texture coordinates on Y axis. This assumes
now that in texture addressing Y = 0.0f refers to the top of the
(i.e. the first scan line) texture and 1.0f refers to the bottom
of the texture (i.e. the last scan line). This conveniently aligns
with the fact that when texture data is uploaded GL assumes the
first scanline in the memory of the image data to be the last
scanline of the image. This basically means that the normal
convention where Y = 1.0f refers to the top of the image will
sample the last scanline of data (i.e the bottom of the image)
unless either the image is flipped when uploaded or the texture
coordinate is changed.

Instead of changing the texture coordinate via device texture matrix
we can just bake this transformation directly in the drawable shapes'
vertex data and thus simplify the code by removing the need for
such a texture matrix and having to perform the matrix mult in the
shader and also simplify the logic of the code (such as the material's
texture box) that needs to compute positions within the texture.

Additionally this scheme also aligns with the way the gl_PointCoord
varies across a point when point fragments are rasterized. So nothing
special needs to be done when using gl_PointCoord either.