Refactor material widget to use material's JSON serialization
Refactor material widget to use material's JSON serialization in
order to simplify and remove the superfluous resource property based
texture map data loading and saving.

TODO: Seems that this change changes behaviour slightly since the
material is now being held as a member. When all textures are deleted
the rendering still renders an image using the last texture. The
rendering however goes to black when the window is resized. I guess
this is because the texture object is still actually around and
probably bound to the texture unit from before. Possibly doing a
window resize makes Qt bind some other texture in the same unit
which then gets deleted and thus invalidates the texture object
and makes the rendering black.