Refactor TextBuffer text and styling setting and add JSON support
Refactor TextBuffer text and styling setting and add JSON support.
This change removes the TextStyle type and combines the styling
information with the Text type structure which is now public in the
class API in order to support easy read access to the current text
buffer text content.

Additionally add support for JSON based serialization.