Add support for graphics resource (file) mapping and lookup
Add support for graphics resource (file) mapping and lookup.
This indirection gives flexibility in terms of looking up resource
files (font, texture and shader files) from the file system. This
is relevant/applicable when resource lookup needs to support:

 * relative file system file paths and runtime path resolution.
 * swapping one resource file for another implicitly.
 * mapping texture content from packed texture files (atlasses).

Currently this only does resource (file name) translation/mapping
but conceivably also the actual loading of content (file data)
could be added to the same mechanism at some later time.

The currently planned use cases for this behaviour are:

* using relative paths in the Editor whenever possible (i.e. file
  can be expressed as a path relative to the current workspace or
  the editor itself).
* loading resource descriptors / packed texture atlasses when
  running a game.

It should be easy to implement a resource packing scheme using
the resource mapping mechanism for textures however it might be
that some caching is needed to prevent continuous texture box
computations and mappings. This needs to be profiled.

In terms of packing materials (and textures) and the texture boxes
the other option would be to have a more complicated packing scheme
where the material JSON is rewritten to reflect the packing of
textures into atlas files. This would remove the need to lookup
texture boxes but would make for more complicated packing.