Implement workspace management functionality
Implement workspace management functionality including opening
existing workspaces, creating new workspaces and closing the current
workspace.

Currently workspace is defined as a "some random directory where
we expect to find content.json and workspace.json". There's no
fancy UI for opening a directory instead only the Qt's file dialog
is used to select and existing directory. This works but is not
necessarily very obvious especialy so in the case of creating a
new workspace. Perhaps a custom dialog with options such as setting
a workspace name and possibly a flag for creating a directory could
be better but that's for later time.

Note that regarding the UX for creating a new workspace there's
a temptation to just create a new workspace object and let the user
start adding content to it. However this does not work with file
mapping (coming later) for using file paths relative to the workspace.
(I.e for that to work we must know where the workspace is to be saved)

Additionally at some point we might want to copy some "initial content"
(such as shaders) over to the new workspace.