Change OpenGL implementation from QOpenGLWidget to QOpenGLWindow
Change OpenGL implementation from QOpenGLWidget to QOpenGLWindow.
This is done because QOpenGLWidget has rather poor performance.
This appears to be related to the way the whole top level window
(TWL) is being composited when it includes a GL widget and other
widgets. The Qt documentation mentions the following:

QOpenGLWindow documentation says:

"Unlike QOpenGLWidget, QOpenGLWindow has no dependency on the
 widgets module and offers better performance."

QOpenGLWidget documentation says:

"Desktop applications that understand the limitations of this
 approach (for example when it comes to overlaps, transparency,
 scroll views and MDI areas), can use QOpenGLWindow with
 QWidget::createWindowContainer(). This is a modern alternative
 to QGLWidget and is faster than QOpenGLWidget due to the lack of
 the additional composition step."

Additional commentary/info about the problem is available in the
following bug report.

https://bugreports.qt.io/browse/QTBUG-54034

"The big bottleneck remains uploading the raster buffer for the
 entire TLW. A TLW containing a very large QOpenGLWidget and a
 one pixel QWidget results in upload of a texture the size of the
 entire TLW."

With the change from QOpenGLWidget to QOpenGLWindow the performance
is slightly better but still nowhere near where it should be.

The new implementation now creates an instance of GfxWindow which
"is-a" QOpenGLWindow. This is then put inside a "container" through a
call QWidget::createWindowContainer. All this happens inside our
GfxWidget (which "is-a" now just a QWidget) implementation in order
to retain the existing GfxWidge interface to the rest of the codebase.
Essentially for most practical purposes the GfxWidget is now just a
simple shim class.

This change also introducs some other changes that follow from the
different behaviour/capabilities of these two classes:

* A timer is no longer used to trigger a repaint. Instead a repaint is
  scheduled on swap buffer signal. This (by default) synchronises to
  vlank and renders at the display refresh frequency.

* A custom context menu can no longer be triggered for changing the
  widget's background color or toggling the FPS printing. This is
  because the GfxWindow now covers the area of the the underlying
  GfxWidget, so it'd never receive the request for the context menu
  and additionally even if it did open a context menu the GfxWindow
  would probably obscure the said menu. (This is now a "real" window
  and doesn't quite follow the stacking order rules the same way
  as other widgets do).

  As a quick fix we simply now react to F1 and F2 keys to toggle
  the FPS display and to promp for the color dialog respectively.

* Capturing mouse wheel events in an eventFilter in order to translate
  mouse scroll and +/- keys into zoom in/out actions no longer works.
  I suspect this is now due to different event passing semantics and
  propagation rules since we're now dealing with an actual window object.

  In order to resolve this issue the functionality to translate mouse
  wheel events and +/- keys into zoom in/out events (callbacks) is now
  implemented in the GfxWidget. The MainWidgets that use the GfxWidget
  and want to do zoom in/out need to now bind to those callbacks too.

* Disposing of graphics resourcs is now explicit throughout. In order
  to avoid obscure bugs related to having lost the underlying OpenGL
  context it's better to make sure that our graphics cleanup is very
  explicit rather than implicit. I.e. before a a window / dialog
  is actually closed we should call dispose on the GfxWidget. This
  will then request the actual GfxWindow to dispose of the graphics
  device (and its GL resouces) and the graphics painter object.
  Failing to do this is considered a BUG and will ASSERT.

TODO: There's now duplicate callback binding, the GfxWidget binds
to the callbacks of the GfxWindow and the actual client of GfxWidget
binds to the callbacks of GfxWidget. Maybe this could be cleaned up
and the callbacks in GfxWindow removed. (For example handle the Qt
event's coming for the window already in the widget, maybe through
an eventFilter?)

References:
https://doc.qt.io/qt-5.12/qopenglwindow.html#details
https://doc.qt.io/qt-5/qopenglwidget.html#details
https://bugreports.qt.io/browse/QTBUG-54034