Enhance the JSON based description of the packed image
Enhance the JSON based description of the packed image. Instead of
having multiple objects have an array of image objects with source
image name as one image field. This is much easier to read back.
Additionally add an index field for reconstructing the same order
the images were in the source input image list.