Fix a potential bug in incorrectly sharing cached text raster buffer
Fix a potential bug in incorrectly sharing cached text raster buffer.
In case there are previous users of a cached raster buffer (bitmap)
we cannot re-use the shared/cached bitmap in which to rasterize and
compose the text since this would overwrite/invalidate the previous
contents. In this case we unfortunately need to allocate a new bitmap
since the caching is simplitic and doesn't allow multiple bitmaps
of same dimensions (i.e. with same key) to cached. However the
expectation is that the returned bitmap is only used to upload to the
GPU and then discarded rather than being held onto so this should not
affect performance.