Trigger animation updates continuously from MainWindow
Trigger animation updates continuously from MainWindow. We cannot
rely on the time step from GfxWidget to update any animation or
simulation state since the time steps can be irregular. When the
GfxWidget is obscured the rendering stops and there are no more
calls to paintGL. Subsequently when the window becomes visible
again the next call to paintGL will yield arbitrarily large time
step. Using this as the time delta for animations might be unstable.

Current solution is to have a high frequency timer in the main window
and trigger updates that way.