Add first version of material gradient color support
Add first version of material gradient color support. The material
now has 4 color values that map to the "corners" of the material,
to the texture coordinates (0,0), (1,0), (0, 1), (1,  1). The
gradient is then modulated by the base color.

TODO: perhaps allow the gradient coordinates to be offset by the
material coordinate velocity. (Optional flag maybe)