Support eventlog size hints for the horizontal scrollbar to work
Support eventlog size hints for the horizontal scrollbar to work.
Not really quite sure what the size hint should be and how
to come up with it, but without the horizontal scrollbar doesn't
appear when needed.

For now use a default font to compute pixel width of the row's text
using QFontMetrics.