Allow a single geometry object to contain multiple draw commands
Allow a single geometry object to contain multiple draw commands
each with their own parameters: type, offset and count.
This will be need for some more complicated shapes that will use
different type of draw for different parts of the shape. For example
rounded rect could use triangles for the body and triangle fan for
the rounded corners.