Add rectangle with rounded corners aka "round rect"
Add rectangle with rounded corners aka "round rect".

Currently the drawable always uses a fixed corner radius. It might
be deseriable to be able to adjust the corner radius. Also extreme
transformations that change the shape from close to a square to
an elongated shape (in either direction) will distort the roundness
of the corners. However currently there's no mechanism to understand
the eventual transformation of the shape as it will appear once
rendered in order to affect the tessellation of the corners.