Add user specific properties to the workspace
Add user specific properties to the workspace. The idea is that the
backing file for these properties can be unique to every user and
would contain data that is only specific to the user's workspace
settings. For example the previously opened widgets.

Save the current list of open widgets on application exit in user
properties instead of workspace properties.