Add support for basic stencil buffer masks in animation drawing
Add support for basic stencil buffer masks in animation drawing,
in other words implement the "Mask" render pass option for
animation nodes.

The draw packets produced by the animation nodes in the animation
are sorted into layers and each layer has two lists. The mask list
and the draw list and for each layer we first render all the shapes
in the mask list in the stencil buffer and then render all the
draw shapes.

TODO:
* improve the visualization of the mask shape in the editor widget.
* maybe support "invert" mask flag / different type of mask.