Refactor pinyin-invaders to draw background using an animation
Refactor pinyin-invaders to draw background using an animation
that is loaded from the game content resources.

This is the first commit that enables the whole content creation
chain from raw assets to content creation in the editor to loading
in the game.

The expected *content* workflow is as follows:

1. The raw game resources (textures) are discovered or created and
   placed in the pinyin-invaders/assets folder.

1.1 The resources can also be placed in Git

2. Using the editor some higher level content is created, for example
   an animations. This content might use and refer to the resources
   assimilated in the assets folder (or elsewhere in the filesystem
   for that matter).

2.2 Create new workspace. This produces content.json and workspace.json
    when the workspace is saved. These files can also be placed in Git.
    If the content is placed so that it's relative to the workspace
    then the content paths stay relative and it's possible to keep the
    workspace completely portable, i.e. be able to open it on any machine
    regardless what is the path where the project has been checked out to.

    If the resource files are in a path that is not rooted in the workspace
    then the paths will be stored as absolute paths and this might make
    the workspace non-portable from one machine to another.

3. Workspace content/resources are packaged for the consumption of the
   game. This does will place the "raw" (file level) resources in the
   output folder and also perform activities such as combining textures
   into texture atlasses and pre-rasterizing content such as text.
   The resources included are texture files (.png, jpg etc), font files
   (.otf, .ttf) and also the shader files (.glsl).

4. The game is launched and it loads the output (content.json) from
   step 3.

5. The game can refer to the content loaded from content.json through
   the ResourceLoader object. (Currently only animations and the
   low level gfx objects, i.e. materials and drawables are available).

The software development workflow is follows:

1. Checkout this project
2. Go into the build folder (see README) and build
3. The build process will build the editor and the games + test apps
   and copy the shaders from graphics/shaders to the installation
   locations of the binaries (the graphics test app and the editor)
4. From the game perspective files such as shaders are *content*
   and are therefore part of the software build system but part of the
   content packaging system. Thus the software build does no longer
   copy the shader files.

Note that currently the game is in a "mixed" mode where some of the
content is being rendered and accessed through the game directly
from the "old" folder layout and part of the content is now going
trough the content workflow mentioned above. This will be refactored
over time as more and more content is moved from being handled by the
game directly into being handled by the editor application and then
being built and deployed in the project packaging step.

Currently it's also possible to launch the game in a "development
mode" using the content.json produced by the workspace when it's saved
since the *format* of these files is the same. This will then use
the content in an "unpackaged" and (unoptimized) form. Right now
it will mean having to start the game in the folder that contains
the content.json (next to workspace.json).